Assign Visit to the Cluster

Assign a given visit to the closest cluster from a given clustering. The function generates the embedding of the visit (by the function embed_visit) and computed the squared Euclidean distance from the clusters' centers.

assign_visit_to_cluster(visit_description, clusters, inter_term_vectors,
  exam_term_vectors)

Arguments

visit_description
A named two element vector (with names: inter and exam) of concatenated terms of interview and examination, separated by ", "

clusters
An output of the function cluster_visits

inter_term_vectors
A matrix of embeddings of the interview terms

exam_term_vectors
A matrix of embeddings of the examination terms

Value

A list of:

cl

The number of the closest cluster

distances

The vector of distances to each cluster

Examples

inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
                                  term_count_min = 1L)

#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 1, expected cost 0.1702
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 2, expected cost 0.0983
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 3, expected cost 0.0628
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 4, expected cost 0.0421
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 5, expected cost 0.0293
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 6, expected cost 0.0209
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 7, expected cost 0.0154
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 8, expected cost 0.0116
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 9, expected cost 0.0088
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 10, expected cost 0.0069
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 11, expected cost 0.0054
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 12, expected cost 0.0042
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 13, expected cost 0.0033
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 14, expected cost 0.0027
#> INFO [2020-05-29 09:12:02] 2020-05-29 09:12:02 - epoch 15, expected cost 0.0021

exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
                                 term_count_min = 1L)

#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:02 - epoch 1, expected cost 0.0729
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 2, expected cost 0.0370
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 3, expected cost 0.0243
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 4, expected cost 0.0180
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 5, expected cost 0.0142
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 6, expected cost 0.0118
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 7, expected cost 0.0100
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 8, expected cost 0.0087
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 9, expected cost 0.0076
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 10, expected cost 0.0068
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 11, expected cost 0.0061
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 12, expected cost 0.0054
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 13, expected cost 0.0049
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 14, expected cost 0.0044
#> INFO [2020-05-29 09:12:03] 2020-05-29 09:12:03 - epoch 15, expected cost 0.0040

visits_vectors <- embed_list_visits(interviews,
                                    examinations, inter_term_vectors, exam_term_vectors)
clusters <- cluster_visits(visits_vectors, visits,
                           spec = "internist",
                           cluster_number = 2L)
assign_visit_to_cluster(c(inter = "cough, fever, rhinitis", exam = ""),
                       clusters, inter_term_vectors, exam_term_vectors)

#> $cl
#> 2 
#> 2 
#> 
#> $distances
#>          1          2 
#> 0.06375225 0.03257228 
#> 

assign_visit_to_cluster(c(inter = "thyroid", exam = ""),
                       clusters, inter_term_vectors, exam_term_vectors)

#> $cl
#> 2 
#> 2 
#> 
#> $distances
#>          1          2 
#> 0.10117841 0.06527791 
#> 

assign_visit_to_cluster(c(inter = "cough, fever, rhinitis",
                         exam = "patient, heart"),
                       clusters, inter_term_vectors, exam_term_vectors)

#> $cl
#> 2 
#> 2 
#> 
#> $distances
#>          1          2 
#> 0.09267911 0.03307892 
#>