Choose Embedded Pairs

Choose embedded pairs for given embeddings of terms and pairs of terms. This function return only the pairs where both terms have embeddings.

.choose_embedded_pairs(term_vectors, terms)

Arguments

term_vectors
A matrix of embeddings of the terms

terms
A list of two character vectors: the first and the second elements of pairs

Value

A list of two character vectors: the first and the second elements of pairs

Examples

inter <- parse_data(interview_1000)
term_vectors <- embed_terms(unique(inter[[1]]$term),
  inter[[2]], embedding_size = 5)

#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 1, expected cost 0.0895
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 2, expected cost 0.0528
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 3, expected cost 0.0438
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 4, expected cost 0.0396
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 5, expected cost 0.0370
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 6, expected cost 0.0351
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 7, expected cost 0.0338
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 8, expected cost 0.0328
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 9, expected cost 0.0321
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 10, expected cost 0.0315
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 11, expected cost 0.0311
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 12, expected cost 0.0307
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 13, expected cost 0.0304
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 14, expected cost 0.0301
#> INFO [2019-12-13 09:21:08] 2019-12-13 09:21:08 - epoch 15, expected cost 0.0299

terms <- terms_pairs_test[[1]]
terms

#> [[1]]
#>  [1] "ból barku"             "ból brzucha"           "ból biodra"           
#>  [4] "ból gardła"            "ból głowy"             "ból grzbietu"         
#>  [7] "ból karku"             "ból klatki piersiowej" "ból kolana"           
#> [10] "ból kręgosłupa"        "ból krzyża"            "ból łokcia"           
#> [13] "ból łydki"             "ból nogi"              "ból oka"              
#> [16] "ból pleców"            "ból stawu"             "ból stopy"            
#> [19] "ból szyi"              "ból ucha"              "ból zęba"             
#> [22] "ból żołądka"          
#> 
#> [[2]]
#>  [1] "bark"             "podbrzusze"       "biodro"           "gardło"          
#>  [5] "głowa"            "grzbiet"          "kark"             "klatka piersiowa"
#>  [9] "kolano"           "kręgosłup"        "krzyż"            "łokieć"          
#> [13] "łydka"            "noga"             "oko"              "plecy"           
#> [17] "staw"             "stopa"            "szyja"            "ucho"            
#> [21] "ząb"              "żołądek"         
#> 

.choose_embedded_pairs(term_vectors, terms)

#> Error in .choose_embedded_pairs(term_vectors, terms): nie udało się znaleźć funkcji '.choose_embedded_pairs'