Generate an Embedding of a Visit

Generate an embedding of a given visit based on interview and examination descriptions and the embeddings of terms. This function calls embed_description function and merge the obtained embeddings of the interview and the examination.

embed_visit(visit_description, inter_term_vectors, exam_term_vectors)

Arguments

visit_description
A named two-element vector (names: "inter" and "exam") of concatenated terms of interview and examination. Terms must be separated by ", ".

inter_term_vectors
A matrix of embeddings of the interview terms

exam_term_vectors
A matrix of embeddings of the examination terms

Value

A vector of the embedding of the visit.

Examples

inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
 term_count_min = 1L)

#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 1, expected cost 0.1702
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 2, expected cost 0.0983
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 3, expected cost 0.0628
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 4, expected cost 0.0421
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 5, expected cost 0.0293
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 6, expected cost 0.0209
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 7, expected cost 0.0154
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 8, expected cost 0.0116
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 9, expected cost 0.0088
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 10, expected cost 0.0069
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 11, expected cost 0.0054
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 12, expected cost 0.0042
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 13, expected cost 0.0033
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 14, expected cost 0.0027
#> INFO [2020-05-29 09:12:05] 2020-05-29 09:12:05 - epoch 15, expected cost 0.0021

exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
  term_count_min = 1L)

#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 1, expected cost 0.0729
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 2, expected cost 0.0370
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 3, expected cost 0.0243
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 4, expected cost 0.0180
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 5, expected cost 0.0142
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 6, expected cost 0.0118
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 7, expected cost 0.0100
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 8, expected cost 0.0087
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 9, expected cost 0.0076
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 10, expected cost 0.0068
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 11, expected cost 0.0061
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 12, expected cost 0.0054
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 13, expected cost 0.0049
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 14, expected cost 0.0044
#> INFO [2020-05-29 09:12:06] 2020-05-29 09:12:06 - epoch 15, expected cost 0.0040

embed_visit(c(inter = "", exam = "heart, woman, eye"),
  inter_term_vectors, exam_term_vectors)

#>  [1]          NA          NA          NA          NA          NA          NA
#>  [7]          NA          NA          NA          NA -0.11948839 -0.01015102
#> [13]  0.03705520 -0.08379906 -0.19850381  0.08107994  0.23177070 -0.21579767
#> [19] -0.02647596  0.01301111

embed_visit(c(inter = "cough",
             exam = "heart, woman, eye"),
           inter_term_vectors, exam_term_vectors)

#>  [1]  0.004778572  0.378802598  0.417534798 -0.478798181  0.160945043
#>  [6] -0.444853812  0.459734917 -0.126606524 -0.179721281 -0.408498943
#> [11] -0.119488393 -0.010151019  0.037055204 -0.083799064 -0.198503807
#> [16]  0.081079945  0.231770699 -0.215797673 -0.026475956  0.013011110

embed_visit(c(inter = "cough", exam = ""),
           inter_term_vectors, exam_term_vectors)

#>  [1]  0.004778572  0.378802598  0.417534798 -0.478798181  0.160945043
#>  [6] -0.444853812  0.459734917 -0.126606524 -0.179721281 -0.408498943
#> [11]           NA           NA           NA           NA           NA
#> [16]           NA           NA           NA           NA           NA

embed_visit(c(inter = "", exam = "objaw chorobowy"),
           inter_term_vectors, exam_term_vectors)

#>  [1] NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA