Merge and Count Terms from Descriptions

Merge and count terms from descriptions of visits given in a list. The function prepare the data to be convenient to compute embeddings.

parse_data(data)

Arguments

data
A list of 2-element lists with elements:

names

An ID of the visit

annot

A data frame with columns:

lemma

A term extracted from the original description of the visit

type

A category of the term assigned by linguists

Value

The list of two elements:

term_table

A data frame with columns:

term

(Elements in this column do not have to be unique)

category

A category of the term

count

A number of occurrences of the term in the data

merged_terms

A named vector of concatenated unique sorted terms for every description, terms are separated by ", ".

Examples

inter <- parse_data(interview_1000)
tab <- inter[[1]]
head(tab[order(tab$count, decreasing = TRUE), ])

#> # A tibble: 6 x 3
#> # Groups:   term [5]
#>   term      category count
#>   <fct>     <fct>    <int>
#> 1 nic       NEG        523
#> 2 ujemny    wCechy     371
#> 3 wieczorem pora       310
#> 4 sód       chm        200
#> 5 sód       NEW        200
#> 6 nie       NEG        181

inter[[2]][1:5]

#>                                                                                                                                           8834656 
#>                                                                                                                   "ból, temperatura ciała, tempo" 
#>                                                                                                                                           8040952 
#>                                                 "bark, ciśnienie, nadgarstek, nasilenie dolegliwości, nic, poziom, sód, staw ręki, tętno, ujemny" 
#>                                                                                                                                           1898742 
#> "apetyt, duży, lek, lękowy, nic, niewydolność, pamięć, poczucie, poprawa stanu psychicznego, prawe oko, skarga, słaby, sód, ujemny, w ciągu dnia" 
#>                                                                                                                                           1919649 
#>                                                                                           "biegunka, dziecko, nie, poprawa, prawe oko, włączenie" 
#>                                                                                                                                           8029957 
#>                                                                                                 "nasilenie, nasilić, ogólny, osłabić, osłabienie" 

exam <- parse_data(examination_1000)
recom <- parse_data(recommendation_1000)