Visualize Embeddings of Visits

Visualize given embeddings of visits of specified specialties of doctors. The visualization is generated on the 2D plane by t-SNE algorithm or by PCA (two main components) and plotted with ggplot. If t-SNE cannot be generated (because for example perplexity is too large for the number of samples), there are plotted two main components from PCA. Points representing the visits can be colored by doctor ID, the first letter of ICD-10 code or by cluster number. The plot can be optionally saved to the given PDF file.

visualize_visit_embeddings(visits_vectors, visit_table, method = "tsne",
  color_by, spec, clusters, n_doctors = 5L, n_codes = 9L,
  save = FALSE, path_to_save)

Arguments

visit_table
A data frame with columns:

visit_id

icd10

(Optional, if color_by == "icd10"): ICD-10 code of the visit

doctor_id

(Optional, if color_by == "doctor"): doctor's ID

specialties

Specialties of doctors, separated by ", "

method
One of "tsne" (default) or "pca" - a method of generating the plot

color_by
A string, one of {"doctor", "icd10", "cluster"}

spec
A specialty to be plotted (optional, required if color_by != "clusters")

clusters
An output of the function cluster_visits (optional, if color_by == "clusters")

n_doctors
A number of doctors to be colored, the rest will be grey (optional, if color_by == "doctor"), default: 5

n_codes
A number of codes to be colored, the rest will be grey (optional, if color_by == "icd10"), default: 9

save
A logical indicating if the plot should be saved to the file

path_to_save
An optional string of the path to the target PDF file

visit_vectors
A matrix of embeddings of visits

Value

A generated plot of embeddings.

Examples

inter_term_vectors <- embed_terms(interview_1000, embedding_size = 10L)

#> Error in gsub(", ", ",", merged_terms): nie znaleziono obiektu 'interview_1000'

exam_term_vectors <- embed_terms(examination_1000, embedding_size = 10L)

#> Error in gsub(", ", ",", merged_terms): nie znaleziono obiektu 'examination_1000'

visits_vectors <- embed_list_visits(interview_1000,
  examination_1000, inter_term_vectors, exam_term_vectors)

#> Error in unique(c(names(inter_descriptions), names(exam_descriptions))): nie znaleziono obiektu 'interview_1000'

visualize_visit_embeddings(visits_vectors, visits_1000, "tsne",
  color_by = "doctor", spec = "medycyna rodzinna", n_doctors = 10)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, "tsne",     color_by = "doctor", spec = "medycyna rodzinna", n_doctors = 10): nie znaleziono obiektu 'visits_vectors'

visualize_visit_embeddings(visits_vectors, visits_1000, "pca",
  color_by = "doctor", spec = "medycyna rodzinna", n_doctors = 9)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, "pca",     color_by = "doctor", spec = "medycyna rodzinna", n_doctors = 9): nie znaleziono obiektu 'visits_vectors'

visualize_visit_embeddings(visits_vectors, visits_1000,
  color_by = "icd10", spec = "medycyna rodzinna", n_codes = 4)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, color_by = "icd10",     spec = "medycyna rodzinna", n_codes = 4): nie znaleziono obiektu 'visits_vectors'

visualize_visit_embeddings(visits_vectors, visits_1000,
  color_by = "icd10", spec = "kardiologia", n_codes = 2)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, color_by = "icd10",     spec = "kardiologia", n_codes = 2): nie znaleziono obiektu 'visits_vectors'

visualize_visit_embeddings(visits_vectors, visits_1000,
  color_by = "doctor", spec = "ginekologia", n_doctors = 5)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, color_by = "doctor",     spec = "ginekologia", n_doctors = 5): nie znaleziono obiektu 'visits_vectors'

clusters <- cluster_visits(visits_vectors, visits_1000,
  spec = "medycyna rodzinna", cluster_number = 5)

#> Error in cluster_visits(visits_vectors, visits_1000, spec = "medycyna rodzinna",     cluster_number = 5): nie znaleziono obiektu 'visits_1000'

visualize_visit_embeddings(visits_vectors, visits_1000,
  color_by = "cluster",
  clusters = clusters)

#> Error in visualize_visit_embeddings(visits_vectors, visits_1000, color_by = "cluster",     clusters = clusters): nie znaleziono obiektu 'visits_vectors'