Generate Embeddings of Visits from a List

Generate embeddings of given visits, contained in two named vectors. One vector contains the descriptions of interviews and the second -- the descriptions of examinations. This function interatively calls embed_visit function.

embed_list_visits(inter_descriptions, exam_descriptions,
  inter_term_vectors, exam_term_vectors)

Arguments

inter_descriptions
A vector of descriptions of interviews, named by visits' IDs; terms in the descriptions are separated by ", "

exam_descriptions
A vector of descriptions of examinations, named by visits' IDs; the visits' IDs do not have to be identical to inter_descriptions

inter_term_vectors
A matrix of embeddings of the interview terms

exam_term_vectors
A matrix of embeddings of the examination terms

Value

A matrix of embeddings of visits -- in each row there is an embedding of one visit.

See also

Examples

inter_term_vectors <- embed_terms(interviews, embedding_size = 10L,
  term_count_min = 1L)

#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 1, expected cost 0.1702
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 2, expected cost 0.0983
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 3, expected cost 0.0628
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 4, expected cost 0.0421
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 5, expected cost 0.0293
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 6, expected cost 0.0209
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 7, expected cost 0.0154
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 8, expected cost 0.0116
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 9, expected cost 0.0088
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 10, expected cost 0.0069
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 11, expected cost 0.0054
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 12, expected cost 0.0042
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 13, expected cost 0.0033
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 14, expected cost 0.0027
#> INFO [2020-06-05 12:42:03] 2020-06-05 12:42:03 - epoch 15, expected cost 0.0021

exam_term_vectors <- embed_terms(examinations, embedding_size = 10L,
  term_count_min = 1L)

#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 1, expected cost 0.0729
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 2, expected cost 0.0370
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 3, expected cost 0.0243
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 4, expected cost 0.0180
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 5, expected cost 0.0142
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 6, expected cost 0.0118
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 7, expected cost 0.0100
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 8, expected cost 0.0087
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 9, expected cost 0.0076
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 10, expected cost 0.0068
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 11, expected cost 0.0061
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 12, expected cost 0.0054
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 13, expected cost 0.0049
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 14, expected cost 0.0044
#> INFO [2020-06-05 12:42:04] 2020-06-05 12:42:04 - epoch 15, expected cost 0.0040

visits_vectors <- embed_list_visits(interviews,
                                    examinations,
inter_term_vectors, exam_term_vectors)
nrow(visits_vectors)

#> [1] 11

visits_vectors

#>            [,1]         [,2]         [,3]         [,4]        [,5]        [,6]
#> 101 -0.01787458  0.236947818 -0.029921770 -0.085469021  0.13657670  0.17326332
#> 102 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 103 -0.26034198  0.003943935 -0.003855810  0.011684724  0.28185698 -0.23633351
#> 104 -0.01787458  0.236947818 -0.029921770 -0.085469021  0.13657670  0.17326332
#> 105 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 106 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 107 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 108 -0.11659506  0.092524713 -0.009171322 -0.099841253 -0.06115377  0.02065507
#> 109 -0.11659506  0.092524713 -0.009171322 -0.099841253 -0.06115377  0.02065507
#> 110 -0.35906245 -0.140479170  0.016894639 -0.002687508  0.08412650 -0.38894176
#> 111          NA           NA           NA           NA          NA          NA
#>           [,7]        [,8]         [,9]       [,10]        [,11]       [,12]
#> 101 0.08520472  0.30467191  0.054711033  0.07044457           NA          NA
#> 102 0.52067906  0.02000736 -0.240970828 -0.38121587 -0.010564044 -0.19152130
#> 103 0.31253360  0.33364742 -0.171810698 -0.34143141  0.003727902  0.13277787
#> 104 0.08520472  0.30467191  0.054711033  0.07044457  0.093647156  0.12660846
#> 105 0.52067906  0.02000736 -0.240970828 -0.38121587 -0.155399621 -0.15491304
#> 106 0.52067906  0.02000736 -0.240970828 -0.38121587  0.014833409  0.14100792
#> 107 0.52067906  0.02000736 -0.240970828 -0.38121587  0.063876513  0.05571919
#> 108 0.10234918 -0.15077619  0.009358175  0.42805733 -0.122630707  0.03959992
#> 109 0.10234918 -0.15077619  0.009358175  0.42805733 -0.090605456  0.13128690
#> 110 0.32967805 -0.12180068 -0.217163555  0.01618135  0.054230120  0.09467864
#> 111         NA          NA           NA          NA -0.065439451  0.10821787
#>            [,13]        [,14]        [,15]        [,16]       [,17]       [,18]
#> 101           NA           NA           NA           NA          NA          NA
#> 102 -0.064329796  0.109247630  0.058476271 -0.077882707 -0.24551799 -0.05784465
#> 103  0.041036060  0.130667954  0.059270306 -0.107490480 -0.10542379  0.22265545
#> 104  0.011021601  0.061413542 -0.072034737 -0.174884200 -0.10762685  0.19446689
#> 105 -0.075609687  0.005400019  0.180959780  0.003463741  0.05443610  0.05864274
#> 106 -0.013135388  0.126385143 -0.302465999 -0.190648686 -0.25231558 -0.21363980
#> 107 -0.009345984 -0.107234917 -0.104841398 -0.173075624 -0.07760018 -0.04975352
#> 108 -0.022584571  0.065401408  0.007187858  0.113845181  0.18888984  0.28423405
#> 109  0.060629157 -0.026927595 -0.164628841 -0.006626189  0.16907078  0.06243548
#> 110  0.071909048  0.076920016 -0.287112350 -0.087972637 -0.13088331 -0.05405190
#> 111 -0.064375961  0.011484621  0.011109995 -0.208357953 -0.19682940 -0.18115286
#>           [,19]       [,20]
#> 101          NA          NA
#> 102 -0.02167543  0.14265516
#> 103 -0.07859598 -0.23821454
#> 104  0.03796901 -0.18084607
#> 105 -0.09097750  0.11219319
#> 106  0.06080274 -0.20023754
#> 107  0.06721641 -0.04968033
#> 108 -0.17684276 -0.19687914
#> 109 -0.01793119 -0.17807858
#> 110  0.05137087 -0.14761662
#> 111 -0.03991671 -0.15966973

visits_vectors[is.na(visits_vectors)] <- 0
D <- as.matrix(dist(visits_vectors))

# Two closest visits
examinations[rownames(which(D == min(D[D > 0]),
                  arr.ind = TRUE))]

#>                                 109                                 108 
#>   "woman, patient, rhinitis, heart" "woman, patient, rhinitis, thyroid" 

visits_vectors <- embed_list_visits(interviews, "",
  inter_term_vectors, exam_term_vectors)
nrow(visits_vectors)

#> [1] 10

visits_vectors

#>            [,1]         [,2]         [,3]         [,4]        [,5]        [,6]
#> 101 -0.01787458  0.236947818 -0.029921770 -0.085469021  0.13657670  0.17326332
#> 102 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 103 -0.26034198  0.003943935 -0.003855810  0.011684724  0.28185698 -0.23633351
#> 104 -0.01787458  0.236947818 -0.029921770 -0.085469021  0.13657670  0.17326332
#> 105 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 106 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 107 -0.03386609 -0.018199205  0.055580780 -0.252045454  0.15427430 -0.46825132
#> 108 -0.11659506  0.092524713 -0.009171322 -0.099841253 -0.06115377  0.02065507
#> 109 -0.11659506  0.092524713 -0.009171322 -0.099841253 -0.06115377  0.02065507
#> 110 -0.35906245 -0.140479170  0.016894639 -0.002687508  0.08412650 -0.38894176
#>           [,7]        [,8]         [,9]       [,10] [,11] [,12] [,13] [,14]
#> 101 0.08520472  0.30467191  0.054711033  0.07044457    NA    NA    NA    NA
#> 102 0.52067906  0.02000736 -0.240970828 -0.38121587    NA    NA    NA    NA
#> 103 0.31253360  0.33364742 -0.171810698 -0.34143141    NA    NA    NA    NA
#> 104 0.08520472  0.30467191  0.054711033  0.07044457    NA    NA    NA    NA
#> 105 0.52067906  0.02000736 -0.240970828 -0.38121587    NA    NA    NA    NA
#> 106 0.52067906  0.02000736 -0.240970828 -0.38121587    NA    NA    NA    NA
#> 107 0.52067906  0.02000736 -0.240970828 -0.38121587    NA    NA    NA    NA
#> 108 0.10234918 -0.15077619  0.009358175  0.42805733    NA    NA    NA    NA
#> 109 0.10234918 -0.15077619  0.009358175  0.42805733    NA    NA    NA    NA
#> 110 0.32967805 -0.12180068 -0.217163555  0.01618135    NA    NA    NA    NA
#>     [,15] [,16] [,17] [,18] [,19] [,20]
#> 101    NA    NA    NA    NA    NA    NA
#> 102    NA    NA    NA    NA    NA    NA
#> 103    NA    NA    NA    NA    NA    NA
#> 104    NA    NA    NA    NA    NA    NA
#> 105    NA    NA    NA    NA    NA    NA
#> 106    NA    NA    NA    NA    NA    NA
#> 107    NA    NA    NA    NA    NA    NA
#> 108    NA    NA    NA    NA    NA    NA
#> 109    NA    NA    NA    NA    NA    NA
#> 110    NA    NA    NA    NA    NA    NA